Project_6_GCP_ETL_Function/requirements.txt
functions-framework==3.*
pandas
pyarrow
google-cloud-storage
google-cloud-bigquery
python-dotenv